Add "copied" indicator
It's nice to know that your copy operation succeeded.  I do change the logo on the form and the favicon when you copy the site password but not when you copy your user name.  This commit adds a "copied!" label that shows up for a second when the copy to clipboard operation succeeds.