Leave scroll bar on
The scroll bar appears when the instructions are open, which makes the main panel shift enough to be annoying.  This commit leaves the scroll bar on all the time.

A better solution would be to attach the scroll bar to the instructions, but I haven't been able to make that work.