Make datalist more visible
A datalist has a white background that doesn't show up well against the white background of the main form.  Unfortunately, there's no way to change the datalist background color or put a border around it.  My solution is to change the background color of the main panel when a datalist is visible.