See settings and their instructions
I normally have only one of help, settings, or instructions open at one time.  However, you might want to see the settings panel while reading the instructions for them.  The problem is that you get horizontal scrolling, which is a no no.  I wonder if this use case is a valid exception.