Got rid of a long standing annoyance
The database was getting a domainname of "".  The problem was that when creating a new entry, I wasn't setting pwdomainname, which is what I set when the password field is in an iframe with a domain name different from that of the page.  I'm now doing that on line 299.

I also improved the notification when you have two, non-identical bookmarks with the same title.