Keep main panel visible while scrolling
I was keeping the main panel visible while you scrolled the instructions if the main panel fit in the popup window.  That wasn't the case when some warning messages were visible.  In those cases, the main panel scrolled off the screen.

This commit keeps the main panel visible even with a large warning message.  Of course, you don't see the whole main panel, but it stops scrolling when either the top or the bottom is on the screen.