Undid a stupid mistake
I added line 174 so that I didn't save the settings if the user said it was phishing, but I used the wrong test.  It isn't phishing if the user says the domain is for the same account, but the new bookmark hasn't been created when this line gets hit.  I mistakenly ran the phishing test again instead of using the cached value.