Open help next to field
I originally had the help at the top of the popup side panel even though Michael suggested that it go next to the relevant field.  I didn't do that due to issues with scrolling.  However, now that I've got that figured out, I can put the help next to the field.