Basic testing infrastructure working
It wasn't easy because the test is doing things that can't happen in normal operation.  For example, the domain name is determined by the active tab or the message sender.  However, testing needs to be able to test what happens at more than one domain name.

Because of that behavior, I need to force both bg.js and ssp.js to throw away any state they remember from the last test.

Now I have to get all the tests to pass.