Keep user provided userid in login form
Before this change, you could type a userid into the page's login form, but it would be replaced by what's in the popup when you clicked the password field.  You'd have to go back and change what's in the user id field.

That doesn't happen with this change. In fact, you can type into the password field, and it won't be changed if you click on the field.