Put warnings to the right of the main panel
There's a fundamental mismatch between the aspect ratio of the popup and that allowed by Chrome.  Putting the warning messages vertically in the main panel exacerbates the mismatch.

This branch puts the messages to the right of the main panel.  That causes the main panel to move, but I want something dramatic like that to draw the user's attention to the warning.

Another issue is settings, instructions, and help.  They appear to the left of the main form, which results in horizontal scrolling if one of them and a message is open.  I think that's probably OK, because a user might want to see the instructions while a warning is open.  Plus, I like that the main panel is always in the middle.