Fix a data race in forget
You get a mouse leave event when you click on the forget button, which results in a siteData message being sent from the popup to the service worker.  Unfortunately, that message contained the settings you are trying to forget.  As a result, the service worker recreated the bookmark it had just deleted.  The fix is to make sure the siteData message has the default settings.