Fix an async error
The new hash function is async, which means generatePassword returns a promise.  The result was an error in bg.js when findpw.js asked for data to fill the login form.  Fortunately, the password doesn't get filled in when loading a page with a login form, so I could just skip the call to generatePassword.