await event handlers
I was using timeouts to make sure async operations had finished before the next step started because event handlers and sendMessage don't return a promise.  It was slow because I was often waiting longer than I needed to.  It was also awkward because I had to accumulate the wait time so the next test didn't start too soon.

This commit uses await for all 3 asynchronous things,  sendMessage, event handlers, and things that return a promise.  I still need sleep to let the triggered event finish, but that's hidden in waitForEvent().  The trick is to wrap the asynchronous thing in a promise and return that.

This commit includes 2 converted tests.  More coming.