Fixed keepAlive logic
Chrome does not clear session storage when the sevice worker shuts down, but Firefox does.  You can't keep the Chrome service worker alive by sending it messages from the content script, but you can with Firefox.  There is no definitive way of knowing which browser from the content script.  Yikes!

You can tell which browser you're in from the service worker, so I have the content script send a message to the service worker every 10 seconds.  If the service worker says it's running in Chrome, the response to the content script tells it to cancel the setInterval.  Whew!

All that wasn't working because I had put the logic in the wrong place.  Instead of a place where it would be hit once, I put it where it would be hit many times, creating the interval timer each time.