Fixed keepAlive bug
The content script sends keepAlive messages to the service worker every 10 seconds if it's running in Firefox but not Chrome.  The service worker tells the content script if it should stop sending them UNLESS there's an error in the messaging.  I forgot that last condition, so keepAlive's kept being sent.