First test passes with promises
I was using setTimeout to wait for event handlers that call async functions to finish.  It worked but had some problems.  First, it was slow, about a minute for 12 tests.  Second, tests would randomly fail, presumably because of the garbage collector.

This test waits for a promise that the event handler creates that it resolves when the async work is done.