All test passing with promises
I was testing async event handlers with setTimeout, which has 2 problems.  It was slow, a minute for 11 tests because the timeouts had to be conservative.  Also, some tests would sometimes fail because the garbage collector made some async events take even longer than the conservative timeout.

This commit wraps event handlers that call async functions or send messages in a promise creator.  Waiting for promises to resolve is much faster, about 15 seconds for the 11 tests.  Also, the garbage collector shouldn't interfere with the tests.