Save partial changes
THIS VERSION DOES NOT WORK because it is doing await on chrome.xxx functions.  I'm going to fix that in branch PGKDF2 and merge the changes into this branch.

This branch simplifies the code in ssp.js to reduce the amount of test related code.  The previous version wrapped each event handler that invokes an async function in a promise.  This version does that wrapping in test.js.  The only thing I have to change in ssp.js is to inclue a resolve statement in the handlers.