Wrap almost all chrome. calls in promises
I was either not waiting for the chrome API calls to finish or I was awaiting them, which is the same as not waiting.  As a result, I had a bunch of race conditions that only showed up in testing.

One solution is continuation passing, which I did in some places, wrapping in promises where that wasn't convenient.