Adjust data-list contrast
When you focus a field with a data-list, site or user name, the data-list appears.  It's sometimes easy to miss it f the list has only one or two entries.  I make it more visible by making "main" darker, but I was making it so dark that it was annoying.  This change makes the background darker by a barely noticeable amount but enough so that the data-list stands out enough not to be easily missed.

This change also resulted in making the background of the side panels white, which isn't a problem.