Fix lost message problem
The service worker doesn't always wake up fast enough to handle the first message, so the message gets lost.  I had previously fixed this problem for a couple of messages but not all of them.  This commit makes sure the service worker is awake by sending "wakeup" messages until the service worker sends an "awake" response.

There is a slight  problem.  There is a noticeable delay when opening the popup or  clicking in the password field when the service worker is idle.  The delay is only around 200 ms.  I think that's related to how long it takes the service worker to restart, so I don't think there's anything I can do about it.