Move warnings to left of main panel
I had them to the right, but seeing the form jump is annoying.  It's supposed to be so you don't miss the warning, but I think it's too jarring.  I tried adding a background color to the warning background, but it didn't fit the look of the other panels.

I also adjusted the width of some items to reduce unneeded white space.