Fixed a problem with provided password
Calculating the password is intentionally slow, but I like to update the sitepw field on every key stroke.  It's not a problem when typing in other fields; the update lags the typing a bit.  The problem comes when you're typing in the sitepw field; the update was overwriting what you just typed.

The solution is to not update the sitepw field if it currently has the focus.