Fix on clipboard indicator and copied! position
If you copy your site password to the clipboard, the logo and tooltip change as an alert.  If you then copy your userid, the alert should clear but it wasn't.

I also moved copied! a few pixels to the right so it wouldn't overlap the label on the site password field.