Fix password export
Now that computing a password takes a minimum of 100 ms, it can take a while to export all your passwords.  The only indication that was happening was the password field changing.  I now have the whole form update as it computes and changed the label on the export passwords button to say Exporting. ...

Another issue was that the popup would close while exporting if you moved your mouse out of the popup.  I couldn't just use autoclose, because message() called in ask2generate would reset it to true.  Instead, I had to change the root element onmouseleave handler.