Change the way I handle mouseout during export
I need to stop handling mouseleave events while exporting passwords or the bookmarks get messed up.  (Yes, I have to fix a lot of mine.)  I was turning off the event handler, but that's a problem if something goes wrong.  Instead I'm setting a flag, exporting.  It's true while exporting and false otherwise and reset to false in a catch block.  Other errors should close the popup, which re-initializes it to false.