Fix to get right site password after export
The settings for each domain name get filled into the form while exporting passwords.  However, the last password is left in the site password field at the end.  That is confusing since the form is filled in with what was there when the export started.  The answer is to just calculate it with the restored settings.