Fix a usability glitch
Say that you have something in the login form password field and change a setting in the popup.  You'd expect the new password to be in the login form, but it wasn't.  Nor was the user id updated if that's what you changed.  The password field is not updated if it is empty, so the normal flow is unchanged.