Got tests working again
There were two problems.  First, I had added a "return true" in the content script's message listener, which resulted in an error that kept the tests from running.  The problem was hard to find because I don't fully understand the error message.

The second problem is that I had removed an event handler that can't be triggered by the user but that I was using for my testing.  I could probably get the same result using an event that the user can generate, but I'll just keep this one for now.