Generate README.md from instructions in HTML
I've been manually editing the files separately even though the content should be the same.  I can simplify things with the following steps.

1. Insert everything in ssp.html from <div id="instructionpanel" to the corresponding </div> into the form at https://www.convertsimple.com/convert-html-to-markdown/.

2. Copy the output to README.md.

3. Use the regular expression

!\[[^\]]*\]\((images\/[^)]*\.(?:png|bmp))\)

and the replacement string

<img src="$1" style="width: 16px; height: 16px; vertical-align: middle;">

to change all occurrences.

4. Manually edit the nested list in Your User Name to insert 4 spaces in the inner list.

The right approach is to create a Github Action to do this, but I haven't figured out how to do that.