Fixed link in help item
I came up with a somewhat too clever way to handle links in the instructions and help that depends on the id of the span starting with the name of the instruction item followed by "ref".  That works fine as long as there's only one of them.  If there's more, then I have to have multiple with the same id, which is an HTML no no.  Fortunately, I don't use getElementByID for these elements, so it doesn't cause any problems.