Fix a bug when clearing superpw on use
I had a sendResponse("persisted") at the end of persistMetadata(), but I need to send a different response when asking for a site password when the "Clear super password on use" is in effect.  As a result, the delivered site password was "persisted" instead of the actual site password.  The fix was to remove the response from the function and put it following the calls that don't have something else to send back.