Some minor code clean up and working on paper
I was getting strange results with special characters and thought I had a bug.  The problem was actually a bad CommonSettings bookmark that I hope was due to something I did during development and isn't actually a bug.  In the process of looking for a non-existent bug I found some code I didn't like.

I'm also updating the paper to reflect recent changes.