Fix algorithm for filtering on strength
I want the returned site password to always have the entropy of a random string of the same length.  My approximate way of doing that wasn't always right.  Fortunately, the zxcvbn report has exactly what I'm looking for.  It tells me if the only viable approach is brute force guessing.