Fix order of operation
There are times when one branch of a conditional has an await and another does not.  If you're not careful, the order of other operations will be different depending on which branch you take.

I had been putting an "await Promise.resolve()" at the start of functions with this problem, but I don't think that's the right way to handle it.  This commit moves the new await into the branches that don't have one.  Sometimes that involves creating an else branch.

Also, in findpw.js I wasn't catching the runtime error check on keepAlive messages, so I moved where I do the check.