Got rid of an undefined
When you blur the special characters, you got an undefined in the settings.  That's because I forgot the second argument to handlekeyup.

The second argument is there so the field name can be different from the element of bg.  I don't actually do that anymore, but I'll leave the option open for the time being.