Let me use old algorithm
Bookmarks modified by the current version properly URIencode special characters, but V2 can't handle that.  As a result, I was unable to get my existing passwords from the V2 extension.  My trick is to have the new version use a different bookmarks folder and restore the old bookmarks from a saved version.  (Wasn't I a clever guy to have done that?)

Of course, I don't want to forget to change it back before deploying V3, so I'll put up with an alert on every page load as a reminder.