Protect super password
By computing for 2-3 seconds create a value that is included in the site password computation.  This computation is done every time you edit the super password.  It increases the adversaries work by a factor of 25.  I'd like at least another factor of 4, but both Doug and Dale says the delay is about as long as they can stand now.