Move settings and instructions
so that they appear between any warnings and the main form.

I originally had them show up to the left of the warning on the theory that the warning was the important element.  It is, but you are probably opening the settings or instructions in response to the warning.  In that case, what you just opened should be close to the main form where you clicked.  It also means what you just clicked to open will appear even if the allowed area is too narrow to show even part of the third column.