Fix datalist
The sitename datalist closed before you got a chance to select an item because moving your mouse into the datalist triggers a mouseleave event on the main panel, which changed the focus to the superpw field.

Removing the change of focus caused other problems, most importantly creating a bookmark with that sitename in spite of an open phishing warning.

There were other minor problems to be dealt with.  For example, the phishing warning stayed open when you selected Pick a new nickname.