Fix auto save issue
Your settings get saved when you mouse out of the main panel.  Unfortunately, that event occurs when you close the settings, instructions, or help panel.  To fix that problem, I set a flag saying not to save settings in those circumstances.

I tried to use the event coordinates instead, but the window changes size when closing any of those panels, so knowing where the click happened is tricky.