Changes for Firefox
Firefox handles datalists differently than Chrome.  The big difference is that Firefox shows the datalist on focus but Chrome does not.  That's only a problem in one scenario, when you click "Pick a new nickname" in the phishing warning.  If you cross the open datalist when mousing to the sitename field before changing the site name, the phising warning opens again.  Of course, you can just type the new site name without moving your mouse, but you'll think there's a problem if you do.  This commit clears the datalist after focusing on the site name field in response to clicking the new nickname button.