Fix an autofill bug
I was autofilling the login form password field in one order of user activities.  If there was something in the login form password field, and the user blurred any of superpw, sitename, or username fields, I would replace the contents of the login form password field with the site password from the popup.

That choice caused two problems.  First, it autofilled a field with potentially new data.  Worse, it replaced what was in the old password field of a change password form.

The fix is to do nothing if there's more than one password field in the form.  If there's only one, then I clear the login form if the popup site password is different from what's in the login form.