Modifed sort by domain name
Most login pages have only one component of the domain name in addition to the base.  For eample, www.apple.com and appleid.apple.com.  If I just sort by domain name, these two entries will be far apart in the SitePassword data file.  If, however, I drop the www and appleid, what's left is apple.com, and the two entries are consecutive.

The strategy doesn't work all the time.  For example, there is also secure4.store.apple.com.  The strategy also fails for twitter.com and x.com.  So the question is if this sort is good enough.