Change handling of multiple password fields
I was requiring you to paste your site password if there was more than one password field on the page.  I didn't want to do that on click because you might click to focus a field of type password that wasn't a password, e.g., a CCV or SSN.

At Doug's prodding, I changed that so you can double click or paste.  That's much more convenient because you don't have to open the popup to get your password inserted.  It's also safer because you don't put your password on the clipboard nearly as often.