Widen area that opens the element menu
I was requiring that the mouse be over the 3 blue dots to open the menu, but I found it annoying that I then had to move the mouse back to the left to get to the icon I wanted.

This commit opens the menu when the mouse reaches the left-most icon.  It feels much more natural.