Dbl-click for user name
Some sites annoy me by forcing me to enter my user name before they show the password field, which means SitePassword can't fill it in for me.  With this change, I can dbl-click in the userid field to get my user name.

If there is no visible password field on the page and only one input field of type "text" or "email", I add a dbl-click handler.  It's a heuristic, so it won't work all the time, and it might add the handler to a field that isn't expecting a user name.

I do not overwrite any pre-existing placeholders or tooltips in case they're telling you something interesting.