First step in using safe suffixes
Some sites have many different domain names for their login pages.  You currently get a phishing warning the first time you see one of them.  It would be nice if I could use the trailing part of the domain name to avoid that, but there's no algorithm for that.

With the new approach you will get a phishing warning only for the second domain name for a given account.  I'll use the common part of the domain names that you earlier said were for the same account to decide if subsequent domain names are phishing.