Added test to common suffixes
to make sure that I still get a phishing warning if there's a common suffix mismatch after creating entries for the same account but different suffixes.

I also renamed "warningbutton" to be "sameacctbutton" because I could never remember what the warning was.

This version also returns the first match when there's potential phishing rather than the last.  It think a better solution will be to return the shortest match, since that's less likely to be an iframe URL.