New different accounts working
I didn't like the way assigning a domain name to a different account worked.  You were changing a site name to change your password and were being asked if you wanted to associate the domain name with a different account.  What's up with that?

The new approach adds an icon to the site name field menu when you want to do that.  It's likely that only a few uses will ever need it, so most people won't be bothered with the concept at all.

There's still work to do: remove the old stuff for handling this, fix the message in the warning, and update the documentation.