Add account icon to site name field
You can change your password for all domain names associated with a given suffix by changing the site name.  If you want to change the password for just one of those domain names, you had to use the account icon in the site password field, which might not be obvious.

This change adds an account icon to the site password field.