UX change to minimizing phishing warnings
The original approach automatically used the settings for a common suffix when encountering a new domain name.  In this version, you have to fill in the site name, but you won't get a phishing warning.  That's better because the site name you pick depends on the page you're looking at.  Also, you won't get trained to click through the phishing warning.