New UX approach works
If you say that two domain names are for the same account, you won't get a phishing warning if you use their site name for other domain names that share the common suffix of the first two.  This version properly decrements the shard suffix count if you forget some of them or delete the corresponding bookmarks.  The safe suffix gets removed if there's only one domain name with it.