Change Related Domain Name button
and some wording.

I've been using red for dangerous actions, green for safe ones, and dark blue for those that give you another try.  The Same Account button on the Related Domain Name warning isn't dangerous, but it's not quite safe either.  Amber seems to be a good compromise between red and green.