Include modified phishing warning
If you use a pre-existing site name, you will get one of two warning.  If the new domain name doesn't share a safe suffix with an old one, you get the full phishing warning.  If it does, you get the modified one.  In the latter case, if you say it's for a different account, you get taken to the form attached to the account button.