Delete safe suffix works
Say that you've said that two domain names with the same suffix are for the same account.  That suffix gets declared safe.  What if you now forget one of those domain names?  Should I keep calling the suffix safe?  What if the user forgot the domain name because it wasn't for the same account?  This commit removes the safe suffix if there's only one domain name in the database that has that suffix.