Fixed a mouse position bug
When you click the same account button in the phishing warning, the warning goes away leaving the mouse outside the popup.  If you don't move the mouse into the popup and then back out, the database doesn't get updated.

Unfortunately, I use the fact that the event doesn't point to any element in the popup to trigger saving the database updates, but the same account event was triggered while the mouse was in the popup. I also need to distinguish from mouseleave events generated when you mouse into a datalist.

My solution is to lie to the mouseleave event by providing null as the event.