Fixed bug in handling duplicate bookmarks
I had changed parseSettings to return the settings instead of an object with the settings, but forgot to change the code where I detect multiple bookmarks with the same title.

I was also opening the alert whether or not the duplicates were identical instead of deleting one of them.

While I was at it, I improved the alert message to explain how to see if the duplicate bookmarks were identical.