Fixed forget problem
If you forget a domain name but don't close the popup, SitePassword remembered your settings.  That might be a good thing if many of them were different from the default, but the behavior is confusing.  You just said Forget, but all your settings for the site name show up.  It's better to start over with the defaults for any site name you enter after Forget.