All tests pass after merge from main branch
Dale reported a bug, so I fixed it, a couple of minor things, and added some tests, all on the main branch as V3.0.11.  I then merged it with this branch.  Most of the conflicts were easy to deal with, but test.js was garbled.  Getting in shape wasn't easy.