Fixed duplicate common settings bkmk problem
I was properly handling duplicate regular bookmarks but not duplicate common settings ones.  The fix was a bit tricky because parsing a common settings bookmark is more complicated than parsing a regular one, which is just the settings.  The common settings one is an object containing a defaultSettings object.  It took a while to find the right place to deal with that because of a poorly named function.