Reload pages with passwords on update
When an extension update is applied, the content script on the page is disconnected from the service worker and popup.  One approach is to reapply the content script, but I get an error.  The other is to reload any page with a password field.

When a page with a password field loads, I remember how many there are on that tab in local storage.  Then I reload the extension those pages when an extension update is applied.

TODO: Remove entries from that list when a tab is closed or you navigate away from the page.