Refresh content script on extension reload
The previous commit was really annoying, because reloading lots of pages is heavyweight and can lose context.  This version reloads the content script instead.

The change wasn't trivial because the old content script is still on the page and gets disconnected from the service worker.  The result is lots of "Extension context invalidated" errors.  They only appear on the developer's console, but it's better to avoid them.

Copilot was useless in finding a fix.  Stackoverflow to the rescue.  It showed some complicated ways to clean things up, but I extracted a very simple fix.  An unloaded content script doesn't have an ID.  Testing that condition lets me skip any operation that will produce an error in an unloaded content script.  It's not clear of Chrome will eventually gargage collect it, but I don't think that's a problem since the content script is so small.