Hide some error messages
I get a "Cannot access contents of the page. Extension manifest must request permission to access the respective host." on a few of my tabs when updating the content script.  The apparent cause is a Content Security Policy at the server, which I can do nothing about.  So, I decided to suppress the message with an errorLogging boolean.