Get rid of errors on update
I now update the content script on all pages when I update the extension.  Unfortunately, the old version continues to run even though it's disconnected.  The result is errors on the developer's console.  Those errors don't seem to have any effect on the behavior of the extension, but it's better to get rid of them.

An earlier commit addressing this problem simply avoided anything that would generate an error in the detached content script but left it running.  This version includes a cleanup function that should allow the detached version to be garbage collected.

One interesting observation is that sometimes the content script starts running before chrome has initialized its runtime.  The result is a reference to a null in something like chrome.runtime.sendMessage.  The solution is to always use chrome.runtime?.sendMessage.