Add developer instructions
When testMode is true, open a web page that explains how to run the tests.

If more than one of the Mode flags is set to true, open a web page that explains that.

I had to change one test that only works on https pages so it will work on the testMode instructions page.