Fix forgotten pwcount
I was having the content script send the count of password fields to the service worker, which the popup would ask for when it opened.  Unfortunately, the service worker forgets it when it is idled.

This version has the popup ask the content script for the count when the popup opens.