Fix nopw warning
I forgot to test my last fix on a page without a password field.  When I did, I didn't get the nopw warning I should have.  The problem was the fix I put in in the e9421c8 commit.  An error in the loop was causing the code to skip to the catch block, so the warning never displayed.

This change takes that loop out, but I'll have to try more web sites to see if I have to put something like it back in.