Remember pwcount across restarts
The service worker forgets its state when it gets idled, but it needs some of it when it wakes back up.  I used to use session storage, but I took it out for a reason I can't remember or find.  I'm putting it back here.

Using session instead of local storage means I don't have to worry about the amount stored.  It's usually a lot less than the number of tabs.

I also needed a better visibility check in findpw.js that Copilot provided.