Fixed page load problem
I didn't see the SitePassword placeholder when I typed a URL into the address bar and hit Enter.  The problem was that the page didn't have focus yet even though loading was complete.  I was using the focus to make sure only one tab at a time was talking to the service worker.  I need a visibility change event handler to deal with that.

My initial attempt ran into a stack overflow because sendpageinfo was calling itself.  Splitting it into 2 functions solved that problem.