Fix same account problem
When I get either of the 2 phishing warnings and select the Same Account button, the settings weren't being saved unless I moused back into the popup before it closed.  The problem was that elementFromPoint was returning the last element the mouse was at because the update of closing the warning panel hadn't settled.

The solution is to run the main panel mouse leave event in the next turn by dispatching it manually.