Fix handling of provided password
When the user provided a password longer than pwlength, I replicated enough of the computed password to make the XOR work.  That choice weakened the encryption like reusing a one-time pad would.  I now extend pwlength to match the provided password in that situation.  I have to reset the default pwlength if you uncheck the provide password box.

User's will have to re-calculate their provided passwords, but I think I'm the only one in that boat.