Attempt to get rid of strange error
I've been getting a strange error about trying to access "query" of a null object in findpw.js.  There is no such reference in that file, so the error must be happening somewhere else and being reported as a broken promise.

The error doesn't appear to affect the exectution of the extension, but it is annoying.