Get rid of another weird error
The page https://www.fastcompany.com/91277240/how-to-spot-fake-job-postings-and-avoid-scams uses overflow for an image tag, which causes Google to issue a warning.  The odd thing is that the warning points to a line in findpw.js where I check an element's properties after being invoked from a mutation observer.

The problem arises because the properties are not evaluated until the element's style is checked, and that line of code is first.  Using a setTimeout in the mutation handler allows a script running on the page to access the style first, which is where the error now points.