Fix not saved problem
Any changes of the settings made when the mouse is outside the popup will be lost if you close the popup without mousing into the popup.  That's because the new values aren't saved until you mouse out.

The fix is easy for the main panel; just send the update to the service worker on keyup.  It's harder for the additional settings because I don't want to calculate the sitepw until you mouse out or blur the element in order to not show error messages while the user is typing.  (Changing pwlength from 12 to 15 involves a step where it is 1.)

I also used the opportunity to clean up the arguments to handleblur.  At one point, I allowed the element id to be different from the name in the settings.  It turns out I didn't need to do that, but I never got around to changing it.