Fix lost superpw problem
I was losing the superpw when the content script sent the onload message to the service worker and when a domain name wasn't in the database.  The problem is that bg is set to the default at those points.  Fortunately, I can just set bg.superpw to the known value.