Remove bookmark for phishing domain name
Since I'm now updating the bookmarks while typing, I can get an entry in the database for a site that turns out to be a phishing site after I finish typing the domain name.  This commit removes that item from the database if I click the green button on the phishing warning.