Move page navigation to bg.js
When the user says it's a real phishing attack, I clear the domain name from the database and tell the content script to clear the login form in bg.js and navigate to chrome://newtab in ssp.js.  That caused a race between the content script message and the page navigation.  The solution is to do them both in bg.js.