Fix saving on each keystroke
I was using a setTimeout to avoid getting a phishing or nopw warning while typing into an input field.  That cause all kinds of strange errors because different keystrokes ran at the same time.  (I didn't think JavaScript worked like that.)

In this commit, I use a flag to say if I want to show a warning for the sitename field.  Other input fields are use handlekeyupnopw(), but I still want to compute sitepw while typing the sitename.