Automatically handle shared credentials
The previous handling of sites that share credentials waited until you filled in the site name.  That depended on your knowing, for example, that disney.com and hulu.com used the same user name and password.  This commit figures that out when the login page loads.  I'm leaving the old handling in place just in case.