Fix strength meter bar
The strength meter bar stayed empty until the strength was > 0.  That might be confusing, so now it shows something as long as there is a value in the password field.

The complication was getting the bar to show the right color because the browser was overriding my settings.  The CSS updates take care of that.