Don't focus any fields when popup opens
A nice usability feature is to focus on the first field the user should fill in.  Unfortunately, that lets the user fill in the form without the mouse ever entering the popup.  Since I can't detect when the popup closes, those changes are lost.  By not focusing on the first field, I force the user to click on the popup.