Undo update database on every keystroke
If you opened the popup and filled in the form without your mouse ever entering the popup, your changes would be lost when the popup closed.  There was no way I could save the updates.

My first try was to update the database on every keystroke, but I kept running into bugs.  My solution now is to not focus any fields when the popup opens.  As result, you are forced to move your mouse into the popup, and I can save the settings when you mouse out.  It was convenient to have an empty field in focus when the popup opens, but it's not worth it.