Fix default focus
In a previous commit I didn't use defaultFocus() when the popup opens to avoid the case of the mouse never entering the popup, leading to lost changes.  However, I found it annoying not having anything in focus.

This commit solves the lost update problem by using defaultFocus() when the mouse enters the popup.  It's not much of a change, but it feels better.  You start moving your mouse to click on a field that gets the focus before you even click.