Fix dbl-click for username
I was sending the getUsername message outside the event handler, which was bad for a couple of reasons.  First, I sent the message on every page update because it's executed inside the mutationObserver.  Second, the page could have been idled between getting the username and dbl-clicking.

I now only send the getName message when dbl-clicking on the input field.